requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
filelock==3.16.1
filetype==1.2.0
fonttools==4.54.1
idna==3.7
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==1.26.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pyparsing==3.2.0
PyQRCode==1.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.48
six==1.16.0
tqdm==4.66.5
urllib3==2.2.3
virtualenv==20.26.5
